// src/types/admin.ts
export interface User {
  _id?: string;
  username: string;
  password: string;
  createdAt: Date;
}